Will now log, check and skip unchanged pages
This should greatly reduce the strain on the webserver as well as
reducing the needed bandwidth bandwidth needed.

Images of a thread are checked against the cache. If all are present in
the cache, the thread will be added to the "last modified" table.

In future, only the threads last-modified attribute in the header will
be checked against the database and used do decide if the page needs to
be parsed again.